Update description in README
In particular, take out the mention of O(n) memory. It's kind of a dumb
thing to say, really. Obviously `sort -u` is going to take O(n) memory
as well. Believe it or not, removing duplicates from an unsorted array
requires that you examine every element! Shocking!

Plus, this was a good opportunity to name-drop Rust a second time, which
shows that I'm one of the cool kids or something.